requirements.txt
autopep8==2.0.0
dnspython==2.2.1
numpy==1.23.5
pandas==1.5.2
pycodestyle==2.10.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
tomli==2.0.1
torch==1.13.1
typing_extensions==4.4.0
